add image register
