+++ src/hooks/useBookRegister.ts
@@ -2,10 +2,17 @@ import { RefObject, useRef, useState } from "react";
 import { useHistory } from "react-router";
 import { collection, addDoc } from "firebase/firestore";
 import { firebaseStore } from "../config/firebase";
+import {
+  getStorage,
+  ref,
+  uploadBytesResumable,
+  getDownloadURL,
+} from "firebase/storage";
 
 type ReturnTypeData = {
   isLoading: boolean;
-  registBookInfo: () => Promise<void>;
+  registerImage: () => void;
+  uploadImage: (e: React.ChangeEvent<HTMLInputElement>) => void;
   titleRef: RefObject<HTMLInputElement>;
   articleRef: RefObject<HTMLInputElement>;
   descriptionRef: RefObject<HTMLTextAreaElement>;
@@ -13,19 +20,44 @@ type ReturnTypeData = {
 
 export const useBookRegister = (): ReturnTypeData => {
   const [isLoading, setLoading] = useState(false);
+  const [image, setImage] = useState<File>();
   const titleRef = useRef<HTMLInputElement>(null);
   const articleRef = useRef<HTMLInputElement>(null);
   const descriptionRef = useRef<HTMLTextAreaElement>(null);
   const history = useHistory();
 
-  const registBookInfo = async () => {
+  const uploadImage = (e: React.ChangeEvent<HTMLInputElement>) => {
+    e.target.files && setImage(e.target.files[0]);
+    console.log('fdsa');
+  }
+
+  const registerImage = async () => {
+    if (image) {
+      const storage = getStorage();
+      const storageRef = ref(storage, `images/${image.name}`);
+      await uploadBytesResumable(storageRef, image)
+        .then((snapshot) => {
+          console.log("Uploaded", snapshot.totalBytes, "bytes.");
+          console.log("File metadata:", snapshot.metadata);
+          getDownloadURL(snapshot.ref).then((url) => {
+            console.log("File available at", url);
+            registBookInfo(url);
+          });
+        })
+        .catch((error) => {
+          console.error("Upload failed", error);
+        });
+    }
+  };
+
+  const registBookInfo = async (url: string) => {
     try {
       setLoading(true);
       const docRef = await addDoc(collection(firebaseStore, "books"), {
         title: titleRef.current?.value,
         article: articleRef.current?.value,
         description: descriptionRef.current?.value,
-        image: "https://source.unsplash.com/random",
+        image: url,
       });
       console.log("Document written with ID: ", docRef.id);
       alert("登録しました");
@@ -36,5 +68,5 @@ export const useBookRegister = (): ReturnTypeData => {
       setLoading(false);
     }
   };
-  return { isLoading, titleRef, articleRef, descriptionRef, registBookInfo };
+  return { isLoading, titleRef, articleRef, descriptionRef, uploadImage,registerImage };
 };
